Add overriden methods equals() and hashCode() for the Grid, create new method addNewLineFilledWithZeros(). Move dataproviders from the GridTest to separate GridData class. Change tests in the PlayfieldTest with regards to Grid usage in the Playfield.
